requirements.txt
# pip install -r requirements.txt

PyYAML
markdown
Jinja2